pyproject.toml
[tool.poetry]
name = "contextuals"
version = "0.1.0"
description = "A library for contextual information support in AI applications"
authors = ["Laurent-Philippe Albou <lpalbou@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/lpalbou/contextuals"
packages = [{include = "contextuals"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
jsonschema = "^4.19.1"
pytz = "^2023.3"
argparse = "^1.4.0"

[tool.poetry.extras]
cli = ["argparse"]

[tool.poetry.scripts]
contextuals = "contextuals.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"